Replace bytes.File by os.File
