+++ stdlib/os/os.go
@@ -86,6 +86,16 @@ func (f *File) Write(b []byte) (int, error) {
 	return n, err
 }
 
+func (f *File) Read(p []byte) (int, error) {
+	//fd := f.innerFile.fd
+	fd := f.Fd()
+	var ptr *byte
+	ptr = &p[0]
+	var nread int
+	nread = read(fd, ptr, cap(p))
+	return nread, nil
+}
+
 func Exit(i int) {
 }
